Convert float to int value

+++ app/Http/Resources/User/UserMinimalItem.php
@@ -40,7 +40,7 @@ class UserMinimalItem extends JsonResource
             'channel' => $this->when($withChannel, $channel),
             'referral_code' => $this->referral_code,
 
-            'loyalty_points' => floatval($this->statistics()->sum('points')),
+            'loyalty_points' => intval($this->statistics()->sum('points')),
         ];
     }
 }
